October 16

+++ Day6/SynthesizerProject/src/main/java/com/example/synthesizerproject/AudioClip.java
@@ -8,20 +8,14 @@ public class AudioClip {
     public static final int sampleRate = 44100;
     public static int TOTAL_SAMPLES = (int) (duration * sampleRate);
 
-    //a member variable that contains the actual byte array
+    //This array will hold the audio data.
     public byte[] byteArray;
 
-    //Methods int getSample( index ) and setSample( index, value ) that return/set
-    //the sample passed as an int. You will need to use bitwise operators to perform these conversions!
-    //The ints that are passed/returned should be in the range of shorts.
-    //These are the closest thing we can do in Java to overloading operator[].
-
-    //Create a constructor?
+    //Create a constructor
     public AudioClip() {
         //Calculate the number of samples
-
         //Initialize data into the array
-        //Multiply it by two to turn the ints into bytes?
+        //Multiply it by two to turn the ints into bytes.
         byteArray = new byte[TOTAL_SAMPLES * 2];
     }
 
@@ -30,21 +24,15 @@ public class AudioClip {
         // Calculate the start index of the 16-bit sample (2 bytes)
         index = index * 2;
         int result = 0;
-        // Combine the lower and upper bytes to form a 16-bit sample
+        // Combine byte1 and byte2 to form a 16-bit sample
         //Originally, these are ints (byte) makes it possible to be used as a byte in lower code
-        byte byte1 = (byte) byteArray[index + 1];
-        byte byte2 = (byte) byteArray[index];
+        byte byte1 = byteArray[index + 1];
+        byte byte2 = byteArray[index];
 
-        //Or the two bytes to get one and return them
+        //Or the two bytes to get one and return it.
         result = byte1 << 8 | (byte2 & 0xFF);
         return result;
 
-//        byte byte1 = byteArray[index*2];
-//        byte byte2 = byteArray[index*2 + 1];
-//        int result = 0;
-//
-//        result = byte2 << 8 | (byte1 & 0xFF);
-//        return result;
     }
 
     public void setSample(int index, int value) {
